microsoft_af/pyproject.toml
[tool.poetry]
name = "microsoft-af"
version = "0.1.0"
description = ""
authors = ["SchwartzKamel <tommy.weber123@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
agent-framework = "^1.0.0b251001"
python-dotenv = "^1.1.1"
agent-framework-devui = "^1.0.0b251007"

[tool.poetry-auto-export]
output = "requirements.txt"
without_hashes = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
